backend/requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
Pillow==11.1.0